+++ app/models/project.rb
@@ -72,8 +72,8 @@ class Project < ActiveRecord::Base
     "//#{Rails.application.config.samson.github.web_url}/#{github_repo}"
   end
 
-  def webhook_stages_for_branch(branch)
-    webhooks.for_branch(branch).map(&:stage)
+  def webhook_stages_for(branch, service_type, service_name)
+    webhooks.for_source(service_type, service_name).for_branch(branch).map(&:stage)
   end
 
   def release_prior_to(release)