+++ db/migrate/20150416231106_add_source_to_webhook.rb
@@ -1,0 +1,12 @@
+class AddSourceToWebhook < ActiveRecord::Migration
+  def up
+    add_column :webhooks, :source, :string, null: false, default: 'any_ci'
+    change_column_default :webhooks, :source, nil
+  end
+
+  def down
+    change_table :webhooks do |t|
+      t.remove :source
+    end
+  end
+end