+++ test/controllers/integrations/jenkins_controller_test.rb
@@ -21,6 +21,6 @@ describe Integrations::JenkinsController do
   before { Deploy.delete_all }
 
   test_regular_commit "Jenkins", no_mapping: {build: { scm: { branch: "foobar" }}}, failed: {build: { status: "FAILURE" }} do
-    project.webhooks.create!(stage: stages(:test_staging), branch: "origin/dev")
+    project.webhooks.create!(stage: stages(:test_staging), branch: "origin/dev", source: 'jenkins')
   end
 end