+++ test/controllers/integrations/semaphore_controller_test.rb
@@ -29,7 +29,7 @@ describe Integrations::SemaphoreController do
   before { Deploy.delete_all }
 
   test_regular_commit "Semaphore", failed: {result: "failed"}, no_mapping: {branch_name: "foobar"} do
-    project.webhooks.create!(stage: stages(:test_staging), branch: "master")
+    project.webhooks.create!(stage: stages(:test_staging), branch: "master", source: 'semaphore')
   end
 
   describe "skipping" do
@@ -56,7 +56,7 @@ describe Integrations::SemaphoreController do
 
     it "doesn't trigger a deploy if we want to skip with [deploy skip]" do
       payload["commit"]["message"] = "[deploy skip]"
-      project.webhooks.create!(stage: stages(:test_staging), branch: "master")
+      project.webhooks.create!(stage: stages(:test_staging), branch: "master", source: 'semaphore')
       post :create, payload.merge(token: project.token)
 
       project.deploys.must_equal []
@@ -64,7 +64,7 @@ describe Integrations::SemaphoreController do
 
     it "doesn't trigger a deploy if we want to skip with [skip deploy]" do
       payload["commit"]["message"] = "[skip deploy]"
-      project.webhooks.create!(stage: stages(:test_staging), branch: "master")
+      project.webhooks.create!(stage: stages(:test_staging), branch: "master", source: 'semaphore')
       post :create, payload.merge(token: project.token)
 
       project.deploys.must_equal []