+++ CHANGELOG.md
@@ -1,1 +1,2 @@
+ - 2015-04-17 [Webhooks] #355 webhooks set to not deploy Github pushes by default
  - 2014-09-03 [Github login] #88 users without organization access cannot login